Add optional systemd build dependency
Maybe because dh-systemd is a transitional in yakkety, the systemd
package, and as a result the systemd.pc package config file, are not
installed in the build environment. Adding a build-dependency breaks
the build on trusty and wily, so adding a dependency on a 'neutral'
package will hopefully make this effectively an optional dependency.